# api/openapi.yaml
openapi: 3.1.0
info:
  title: ElaMachan API
  version: 1.0.0
  description: |
    Single source of truth for the ElaMachan (trilingual Sri Lankan classifieds)
    HTTP contract, per **ADR 0003**. Both the Next.js frontend and the Go backend
    derive from this file:

    - **Frontend** generates a typed TS client/types from this spec
      (`openapi-typescript` + `openapi-fetch`) — no hand-written request types.
    - **Backend** treats this file as the authoritative contract; request/response
      shapes must match it.

    Conventions locked here (do not change without a contract change → **Board
    escalation** per the merge policy):

    - **Versioning:** every endpoint lives under `/api/v1` (the `servers` base).
    - **Error envelope:** all non-2xx responses use `{ "error": { "code", "message" } }`.
    - **Trilingual content:** user/taxonomy text carries `en` / `si` / `ta`
      variants (ADR 0001); the seller authors in one `content_language`.
    - **Auth:** phone + OTP primary (ADR 0002); protected endpoints take a bearer
      session token.

    Status of this v1 surface: `POST /listings/ai-draft` is **implemented**
    (VER-58). The remaining paths (auth, listings, categories, search) are
    **stubs** that lock the shape; each is finalized by its own feature issue and
    is marked `x-status: stub` below.
servers:
  - url: /api/v1
    description: Version 1. Relative to the deployed host (locked `/api/v1` prefix).
...
paths:
...
  /categories:
    get:
      tags: [categories]
      operationId: listCategories
      summary: List the localized category taxonomy
      x-status: stub
      description: |
        **Stub** — finalized by the taxonomy issue (ADR 0001). Returns the
        category tree with names resolved to the requested language via the
        `requested → en → any` fallback chain.
      parameters:
        - $ref: "#/components/parameters/LangQuery"
      responses:
        "200":
          description: The category list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /listings:
    get:
      tags: [listings]
      operationId: listListings
      summary: List published listings
      description: "Paginated browse/filter. Pass mine=true (auth required) to return only the caller's listings."
      parameters:
        - $ref: "#/components/parameters/LangQuery"
        - name: category
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/CategorySlug"
        - name: mine
          in: query
          required: false
          description: "When true, return only the authenticated caller's listings. Requires a valid bearer token."
          schema:
            type: boolean
            default: false
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
      responses:
        "200":
          description: A page of listing summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [listings]
      operationId: createListing
      summary: Create a listing
      description: |
        Creates a listing authored in a single `content_language`. Requires authentication.
        A draft from `/listings/ai-draft` is never auto-created; the seller submits here.
        A PostingPolicy is evaluated first; exceeding the daily cap returns 403.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingCreate"
      responses:
        "201":
          description: Listing created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Posting denied by policy (e.g. daily cap exceeded).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostingDenied"
  /listings/{id}:
    get:
      tags: [listings]
      operationId: getListing
      summary: Get a single listing
      description: "Returns one listing with title/description resolved to the requested language (ADR 0001 fallback)."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/LangQuery"
      responses:
        "200":
          description: The listing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [listings]
      operationId: updateListing
      summary: Update a listing
      description: |
        Edits the caller's own listing. Requires authentication and ownership:
        non-owners get 403, an unknown listing gets 404. `content_language` is
        immutable, so title/description edits land on the listing's existing
        authored-language translation.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: Listing updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [listings]
      operationId: deleteListing
      summary: Delete a listing
      description: |
        Removes the caller's own listing (soft delete; it leaves the catalog and
        dashboard). Requires authentication and ownership: non-owners get 403, an
        unknown listing gets 404.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Listing deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  "/listings/{id}/images:presign":
    post:
      tags: [listings]
      operationId: presignListingImage
      summary: Presign a listing image upload
      description: |
        Owner-only. Validates content-type, size, and the per-listing image cap,
        then returns a short-lived presigned PUT URL the client uses to upload
        the file bytes directly to object storage (no bytes through the API).
        After the PUT succeeds, call `:confirm` to activate the image. Limits:
        at most 8 images per listing; jpeg/png/webp; <= 8MB.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImagePresignRequest"
      responses:
        "201":
          description: Presigned upload created with a pending image record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagePresignResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          description: Image storage is not configured on this server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/listings/{id}/images:confirm":
    post:
      tags: [listings]
      operationId: confirmListingImage
      summary: Confirm a listing image upload
      description: |
        Owner-only. HEAD-verifies the uploaded object exists in storage, then
        activates the pending image record so it is served in listing responses.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageConfirmRequest"
      responses:
        "200":
          description: Image activated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingImage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          description: Image storage is not configured on this server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/listings/{id}/images/{imageId}":
    delete:
      tags: [listings]
      operationId: deleteListingImage
      summary: Delete a listing image
      description: |
        Owner-only. Removes the image object from storage and its record. Used by
        the edit flow.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: imageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Image deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: Image storage is not configured on this server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Session token issued by `POST /auth/otp/verify` (ADR 0002). The exact
        token format (signed JWT vs opaque session) is finalized in the auth
        implementation task.
  parameters:
    LangQuery:
      name: lang
      in: query
      required: false
      description: "Preferred response language; falls back `requested → en → any` (ADR 0001)."
      schema:
        $ref: "#/components/schemas/Lang"
    PageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    Unauthorized:
      description: Authentication is required or failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    CategorySlug:
      type: string
      description: |
        Closed category taxonomy slug. Single source of truth is the backend
        `Categories` set (server-re-validated on AI drafts). Extended via the
        taxonomy issue, not ad hoc.
      enum:
        - electronics
        - vehicles
        - property
        - home_garden
        - fashion
        - mobile_phones
        - services
        - jobs
        - pets
        - other
...
    Category:
      type: object
      description: A taxonomy category with its name resolved to the requested language.
      required: [slug, name]
      properties:
        slug:
          $ref: "#/components/schemas/CategorySlug"
        name:
          type: string
          description: Localized display name (fallback-resolved).
        parent_slug:
          type: [string, "null"]
        sort_order:
          type: integer
      additionalProperties: false
...